Crowbar 0.35 - 2016-09-26_0.35
What's New

=== Decompile ===
New "DeclareSequence QCI File" option that writes a QCI file with a $DeclareSequence line for each sequence in the MDL file. This is useful in the setting up of animations to be used by a different model so that the animations are in the correct order for multiplayer.
Decompile MDL version 45 and 46.

=== Compile ===
FIX: Trying to compile with QC file or folder input set to a root folder like "C:\Test" shows an error message instead of compiling.
FIX: Trying to compile folder or folder + subfolders shows an error message instead of compiling.

=== Set Up Games ===
New options for setting model viewer and unpacker/packer tools.